Warn Device info from Acc DB
To check if Angle is an issue, we are finding the total number of accidents (at single locations) with (Rail->HW  and moving vehicle) and (HW->Rail with the accident on a car under 3). To ensure the correct warning devices are used, warning device information is obtained from the accident database. The inventory database is the latest database and has updated information.